flask-rss/requirements.txt
flask
python-decouple
flask-restful
requests
beautifulsoup4
feedparser